Added Unit Declaration
Added the declaring of "unit" in the constructor.
Added setUnit, can only be called before the onReadyToOperate. There may be times when you want to change it after the sensor is ready to operator. But for now, once node is ready, its set. 
Have not tested the code. 
Good work Tim